#8 Refactor: HID parent and Joystick subclass to make room for remote.
