+++ joystick.h
@@ -9,16 +9,10 @@
 #ifndef JOYSTICK_H_
 #define JOYSTICK_H_
 
-class Joystick {
+#include "hid.h"
+
+class Joystick : public HID {
 private:
-    enum Event {
-        EVENT_NONE  = 0,
-        EVENT_UP    = 0b00001,
-        EVENT_DOWN  = 0b00010,
-        EVENT_LEFT  = 0b00100,
-        EVENT_RIGHT = 0b01000,
-        EVENT_CLICK = 0b10000,
-    };
     static const int sensitivity = 7; // lower is higher :)
 protected:
     int sw_pin, x_pin, y_pin;
@@ -28,23 +22,8 @@ public:
     static const int range = (1024/2) >> sensitivity;
     int autorepeat_delay;
     int autorepeat_repeat;
-    Joystick(int sw_pin, int x_pin, int y_pin);
     unsigned read(void);
-    static bool isEventUp(unsigned event){
-        return (event & EVENT_UP);
-    };
-    static bool isEventDown(unsigned event){
-        return (event & EVENT_DOWN);
-    };
-    static bool isEventLeft(unsigned event){
-        return (event & EVENT_LEFT);
-    };
-    static bool isEventRight(unsigned event){
-        return (event & EVENT_RIGHT);
-    };
-    static bool isEventClick(unsigned event){
-        return (event & EVENT_CLICK);
-    };
+    Joystick(int sw_pin, int x_pin, int y_pin);
     int getPositionX(void);
     int getPositionY(void);
     bool getButtonState(void);
